requirements.txt
python-telegram-bot==20.7
httpx==0.27.0
beautifulsoup4==4.12.3
lxml==5.1.0
supabase==2.4.0
upstash-redis==1.1.0
apscheduler==3.10.4
fake-useragent==1.5.1
python-dotenv==1.0.1
sentry-sdk==1.45.0
